backend/scraper.py
import requests
from bs4 import BeautifulSoup
import json
import os
from datetime import datetime
from typing import Dict, Optional
import time

class EarningsTranscriptScraper:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        self.companies = {
            "AAPL": "Apple Inc.",
            "MSFT": "Microsoft Corporation",
            "GOOGL": "Alphabet Inc.",
            "AMZN": "Amazon.com Inc.",
            "META": "Meta Platforms Inc.",
            "TSLA": "Tesla Inc.",
            "NVDA": "NVIDIA Corporation",
            "JPM": "JPMorgan Chase & Co.",
            "JNJ": "Johnson & Johnson",
            "WMT": "Walmart Inc."
        }
    
    def scrape_transcript(self, ticker: str) -> Optional[Dict]:
        """
        Scrape earnings transcript for a given ticker.
        Note: This is a placeholder that returns mock data.
        In production, you would integrate with actual data sources.
        """
        mock_transcripts = {
            "AAPL": {
                "ticker": "AAPL",
                "company": "Apple Inc.",
                "quarter": "Q1 2025",
                "date": "2025-01-25",
                "content": """
                Apple Q1 2025 Earnings Call Transcript
                
                CEO Tim Cook: Good afternoon everyone. We're pleased to report another strong quarter 
                with revenue of $94.8 billion, up 5% year over year. iPhone continues to show 
                resilience with Services reaching an all-time high.
                
                CFO Luca Maestri: EPS came in at $1.36, beating consensus. We're seeing strong 
                momentum in emerging markets, particularly India and Southeast Asia. Gross margin 
                improved to 45.2%, driven by favorable mix and operational efficiency.
                
                Guidance: For Q2, we expect revenue to be slightly above Q1 levels, with continued 
                strength in Services offsetting seasonal iPhone weakness. We remain confident in our 
                long-term growth trajectory.
                
                Key risks include: Macroeconomic uncertainty, China market dynamics, and supply 
                chain complexities. However, our diversified product portfolio and strong ecosystem 
                position us well for continued success.
                """
            },
            "MSFT": {
                "ticker": "MSFT",
                "company": "Microsoft Corporation",
                "quarter": "Q1 2025",
                "date": "2025-01-24",
                "content": """
                Microsoft Q1 2025 Earnings Call Transcript
                
                CEO Satya Nadella: We delivered another quarter of strong results, with revenue of 
                $62.0 billion, up 12% year over year. Azure growth accelerated to 28%, and our 
                AI initiatives are gaining significant traction.
                
                CFO Amy Hood: Operating income increased 15% to $27.0 billion. We're seeing 
                exceptional demand for our AI and cloud services. Commercial bookings grew 20% 
                with strong multi-year commitments.
                
                Guidance: We expect double-digit revenue growth to continue in Q2, driven by Azure 
                and Office 365. AI services revenue is expected to reach $2 billion run rate by 
                year end.
                
                Opportunities: Massive AI transformation opportunity across all industries. 
                Challenges include: Capacity constraints in AI infrastructure and competitive 
                pressures in cloud market.
                """
            }
        }
        
        if ticker in mock_transcripts:
            time.sleep(1)  # Simulate API delay
            return mock_transcripts[ticker]
        
        return {
            "ticker": ticker,
            "company": self.companies.get(ticker, ticker),
            "quarter": "Q1 2025",
            "date": datetime.now().strftime("%Y-%m-%d"),
            "content": f"Mock earnings transcript for {ticker}. In production, this would contain actual transcript data."
        }
    
    def save_transcript(self, ticker: str, data: Dict):
        """Save transcript data to JSON file"""
        os.makedirs("../data/transcripts", exist_ok=True)
        filepath = f"../data/transcripts/{ticker}_latest.json"
        
        with open(filepath, 'w') as f:
            json.dump(data, f, indent=2)
        
        print(f"Saved transcript for {ticker} to {filepath}")
    
    def scrape_all_companies(self):
        """Scrape transcripts for all configured companies"""
        for ticker in self.companies.keys():
            print(f"Scraping {ticker}...")
            transcript_data = self.scrape_transcript(ticker)
            
            if transcript_data:
                self.save_transcript(ticker, transcript_data)
            else:
                print(f"Failed to scrape {ticker}")
            
            time.sleep(2)  # Rate limiting

if __name__ == "__main__":
    scraper = EarningsTranscriptScraper()
    scraper.scrape_all_companies()