backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.59.0
openai==1.3.0
httpx==0.25.2
yfinance==0.2.33
lxml==4.9.3
pandas==2.1.3
html5lib==1.1
python-multipart==0.0.6